// packages/entities/entities-plugins/src/components/free-form/plugins/datakit/types.ts
export interface FieldMeta {
  name: FieldName
  // type: 'any' | 'string' | 'number' | 'map' | 'object'
}

export interface IOMeta {
  fields: FieldMeta[]
  configurable?: boolean
}

export interface BranchMeta {
  name: BranchName
}

export interface NextMeta {
  branches: BranchMeta[]
}

export interface NodeColors {
  foreground: string
  background: string
}

export interface NodeVisual {
  icon: Component
  colors?: NodeColors
}

export interface NodeMeta extends Partial<NodeVisual> {
  type: NodeType
  summary?: string
  description?: string
  io?: {
    input?: IOMeta
    output?: IOMeta
    next?: NextMeta
  }
  /** hidden in UI */
  hidden?: boolean
}

export type ConfigNodeGroup =
  | 'external_interaction'
  | 'control_flow'
  | 'data_transformation'
  | 'data_value'
  | 'authentication'

export interface ConfigNodeMeta extends NodeMeta {
  type: ConfigNodeType
  summary: string
  description: string
  icon: Component
  group: ConfigNodeGroup
}

export interface ConfigNodeGroupMeta {
  id: ConfigNodeGroup
  title: string
  colors: NodeColors
  nodeTypes: readonly ConfigNodeType[]
}

export interface ConfigNodePanelGroup extends ConfigNodeGroupMeta {
  nodes: readonly ConfigNodeMeta[]
}
// ...
export type NodePhase = 'request' | 'response'

/************************************************
 *             Editor global store              *
 ************************************************/

export type NodeId = `node:${number}`
export type EdgeId = `edge:${number}`
export type FieldId = `field:${number}`
export type GroupId = `${NodeId}:${BranchName}`
export type GroupName = `${NodeName}:${BranchName}`

export type IdConnection = NodeId | `${NodeId}.${FieldId}`

export interface NodeField {
  id: FieldId
  name: FieldName
}

export type UINode = {
  name: NodeName
  phase: NodePhase
  position: XYPosition
  fields: { input?: FieldName[], output?: FieldName[] }
  expanded: { input?: boolean, output?: boolean }
  /** hidden in UI */
  hidden?: boolean
}

export interface NodeInstance {
  id: NodeId
  type: NodeType
  name: NodeName
  phase: NodePhase
  position: XYPosition
  expanded: { input?: boolean, output?: boolean }
  fields: { input: NodeField[], output: NodeField[] }
  /** hidden in UI */
  hidden?: boolean
  config?: Record<string, unknown>
}

export interface ConfigEdge {
  source: NodeName
  sourceField?: FieldName
  target: NodeName
  targetField?: FieldName
}

export interface EdgeData {
  source: NodeId
  sourceField?: FieldId
  target: NodeId
  targetField?: FieldId
}

export interface EdgeInstance extends EdgeData {
  id: EdgeId
}

export interface GroupInstance {
  id: GroupId
  ownerId: NodeId
  branch: BranchName
  phase: NodePhase
  memberIds: NodeId[]
  position?: XYPosition
  dimensions?: Dimensions
}